+++ src/task/mod.rs
@@ -21,12 +21,12 @@
 ****************************************************************************/
 
 pub use self::conv_shapes::convert_shapes_to_paths;
-pub use self::final_fixes::{final_fixes, fix_xmlns_attribute};
+pub use self::final_fixes::{remove_empty_defs, fix_xmlns_attribute};
 pub use self::fix_attrs::fix_invalid_attributes;
+pub use self::group_by_style::group_by_style;
 pub use self::group_defs::group_defs;
 pub use self::join_style_attrs::join_style_attributes;
 pub use self::merge_gradients::merge_gradients;
-pub use self::move_styles_to_group::move_styles_to_group;
 pub use self::preclean_checks::preclean_checks;
 pub use self::resolve_attrs::resolve_attributes;
 pub use self::resolve_inherit::resolve_inherit;
@@ -36,7 +36,10 @@ pub use self::rm_dupl_defs::{
     remove_dupl_radial_gradients,
     remove_dupl_fe_gaussian_blur,
 };
+pub use self::regroup_gradient_stops::regroup_gradient_stops;
+pub use self::resolve_use::resolve_use;
 pub use self::rm_elems::remove_element;
+pub use self::rm_gradient_attrs::remove_gradient_attributes;
 pub use self::rm_invalid_stops::remove_invalid_stops;
 pub use self::rm_invisible_elements::remove_invisible_elements;
 pub use self::rm_needless_attrs::remove_needless_attributes;
@@ -45,22 +48,19 @@ pub use self::rm_unref_ids::remove_unreferenced_ids;
 pub use self::rm_unused_coords::remove_unused_coordinates;
 pub use self::rm_unused_defs::remove_unused_defs;
 pub use self::rm_version::remove_version;
-pub use self::resolve_use::resolve_use;
 pub use self::trim_ids::trim_ids;
-pub use self::ungroup_groups::ungroup_groups;
 pub use self::ungroup_defs::ungroup_defs;
-pub use self::regroup_gradient_stops::regroup_gradient_stops;
-pub use self::rm_gradient_attrs::remove_gradient_attributes;
+pub use self::ungroup_groups::ungroup_groups;
 
 #[macro_use]
 mod macros;
 mod conv_shapes;
 mod final_fixes;
 mod fix_attrs;
+mod group_by_style;
 mod group_defs;
 mod join_style_attrs;
 mod merge_gradients;
-mod move_styles_to_group;
 mod preclean_checks;
 mod regroup_gradient_stops;
 mod resolve_attrs;